[MIG] container_meals: Migration to 16.0
