+++ container_meals/README.rst
@@ -17,7 +17,7 @@ Container Meals
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-coopiteasy%2Faddons-lightgray.png?logo=github
-    :target: https://github.com/coopiteasy/addons/tree/14.0/container_meals
+    :target: https://github.com/coopiteasy/addons/tree/16.0/container_meals
     :alt: coopiteasy/addons
 
 |badge1| |badge2| |badge3|
@@ -70,6 +70,8 @@ Known issues / Roadmap
 
 - Support differing units of measurement.
 - Better support for changing the deposit product.
+- Fix the fact that invoices might use the same deposit multiple times if 
+  created in series before all being payed
 
 Changelog
 =========
@@ -89,7 +91,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/coopiteasy/addons/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/coopiteasy/addons/issues/new?body=module:%20container_meals%0Aversion:%2014.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/coopiteasy/addons/issues/new?body=module:%20container_meals%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -111,6 +113,6 @@ Contributors
 Maintainers
 ~~~~~~~~~~~
 
-This module is part of the `coopiteasy/addons <https://github.com/coopiteasy/addons/tree/14.0/container_meals>`_ project on GitHub.
+This module is part of the `coopiteasy/addons <https://github.com/coopiteasy/addons/tree/16.0/container_meals>`_ project on GitHub.
 
 You are welcome to contribute.
