+++ container_meals/__init__.py
@@ -1,2 +1,1 @@
-from . import controllers
 from . import models
