+++ container_meals/models/__init__.py
@@ -1,5 +1,7 @@
 from . import ir_config_parameter
-from . import product
+from . import product_product
+from . import product_template
 from . import res_config_settings
 from . import res_partner
 from . import sale_order
+from . import sale_order_line
