+++ container_meals/readme/ROADMAP.rst
@@ -1,2 +1,4 @@
 - Support differing units of measurement.
 - Better support for changing the deposit product.
+- Fix the fact that invoices might use the same deposit multiple times if 
+  created in series before all being payed
