+++ container_meals/views/sale_order_views.xml
@@ -1,69 +1,5 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
-    <record id="view_order_line_tree" model="ir.ui.view">
-        <field name="name">container_meals.view_order_line_tree</field>
-        <field name="model">sale.order.line</field>
-        <field name="inherit_id" ref="sale.view_order_line_tree" />
-        <field name="arch" type="xml">
-            <field name="qty_to_invoice" position="after">
-                <!-- TODO: Make this invisible sometimes, maybe -->
-                <field name="not_returned" />
-            </field>
-        </field>
-    </record>
-
-    <record id="sale_order_line_view_tree_not_returned" model="ir.ui.view">
-        <field
-            name="name"
-        >container_meals.sale_order_line_view_tree_not_returned</field>
-        <field name="model">sale.order.line</field>
-        <field name="priority">100</field>
-        <field name="arch" type="xml">
-            <tree string="Sales Order Lines" create="false" editable="top">
-                <field name="order_id" readonly="1" />
-                <field name="order_partner_id" readonly="1" />
-                <field name="name" readonly="1" />
-                <field name="product_uom_qty" string="Qty" readonly="1" />
-                <field name="not_returned" />
-                <field name="product_uom_category_id" invisible="1" />
-                <field
-                    name="product_uom"
-                    string="Unit of Measure"
-                    groups="uom.group_uom"
-                />
-                <field
-                    name="price_unit"
-                    sum="Unit Price"
-                    widget="monetary"
-                    readonly="1"
-                />
-                <field
-                    name="price_subtotal"
-                    sum="Total"
-                    widget="monetary"
-                    readonly="1"
-                />
-                <field name="currency_id" invisible="1" />
-            </tree>
-        </field>
-    </record>
-
-    <record id="sale_order_line_view_form_readonly" model="ir.ui.view">
-        <field name="name">container_meals.sale_order_line_view_form_readonly</field>
-        <field name="model">sale.order.line</field>
-        <field name="inherit_id" ref="sale.sale_order_line_view_form_readonly" />
-        <field name="arch" type="xml">
-            <field name="qty_invoiced" position="after">
-                <field name="is_container" invisible="1" />
-                <field
-                    name="not_returned"
-                    readonly="1"
-                    attrs="{'invisible': [('is_container', '=', False)]}"
-                />
-            </field>
-        </field>
-    </record>
-
     <record id="view_order_form" model="ir.ui.view">
         <field name="name">container_meals.view_order_form</field>
         <field name="model">sale.order</field>
@@ -103,4 +39,5 @@
             </xpath>
         </field>
     </record>
+
 </odoo>
